// src/app/Modules/carrera/model/cpparticipacio.dto.ts
// Generated by https://quicktype.io

import { CpConvocatoriaDTO } from './cpconvocatoria.dto';
import { CpNivell } from './cpnivell.dto';
import { CpPersonalDTO } from './cppersonal.dto';

export interface CpParticipacioDTO {
  codiempl: number;
  convocatoria: CpConvocatoriaDTO;
  dataRegistre: string;
  diesant: number;
  diesit: number;
  diesitant: number;
  diestrebaju: number;
  estat: string;
  estatstr: string;
  grup: string;
  kconvocatoria: number;
  kmotiuexclusio?: number;
  knivellObtingut: number;
  knivellReconegut?: number;
  kparticipacio: number;
  motiuexclusio: string;
  personal: CpPersonalDTO;
  restaant: number;
  swadmes: string;
  swadmesstr: string;
  swdispo4: string;
  totaldies: number;
  computtotaldies: number;
  nivellobtingut?: CpNivell;
  nivellreconegut?: CpNivell;
  questionari: string;
}