// src/app/Modules/carrera/model/cpcursaportat.dto.ts
export interface CpCursAportatDTO {
    data:          string;
    descgp:        string;
    descripcio:    string;
    hores:         number;
    kcurs?:        number;
    kcursaportat:  number;
    kparticipacio: number;
    ordre:         number;
    origen:        string;
    punts:         number;
    swseleccionat: string;
    tipus:         string;
    strconv:       string;
}